// src/app/core/models/product/product.model.ts
export interface Product {
  createdBy: string;
  createdOn: string;
  headOfAccId: number;
  headOfAccName: string;
  modifiedOn: Date;
  modifiedby: string;
  prdBrndId: number;
  prdCatgId: number;
  prdClosingStock: number;
  prdCode: string;
  prdDescription: string;
  prdGrpId: number;
  prdGstPct: number;
  prdHsnCode: number;
  prdMinQty: number;
  prdPurchasedPrice: number;
  prdStatus: number;
  prdTotalValue: number;
  prdUnit: number;
  prdbrndName: string;
  prdcatgName: string;
  prdgrpName: string;
  prdmdlName: string;
  productId: number;
  productbrand: number;
  stockUpdate: Date;
}